adding "UI"

+++ src/main/java/com/griddynamics/user/services/AddressService.java
@@ -3,10 +3,13 @@ package com.griddynamics.user.services;
 import com.griddynamics.user.dtos.AddressDto;
 import com.griddynamics.user.mappers.AddressDtoMapper;
 import com.griddynamics.user.models.Address;
+import com.griddynamics.user.models.User;
 import com.griddynamics.user.repositories.AddressRepository;
 import com.griddynamics.user.repositories.UserRepository;
 import lombok.AllArgsConstructor;
 
+import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.List;
 
@@ -16,28 +19,51 @@ public class AddressService {
     private final AddressDtoMapper addressDtoMapper;
     private final UserRepository userRepository;
 
-    public void addAddress(Long userId, AddressDto addressDto) {
+
+    public boolean addAddress(Long userId, AddressDto addressDto) {
         Address address = addressDtoMapper.addressDtoToAddress(addressDto);
-        addressRepository.save(userId, address);
+        if (userRepository.isUserInDatabase(userId)) {
+            addressRepository.save(userId, address);
+            return true;
+        } else {
+            return false;
+        }
     }
 
-    public void updateAddress(Long userId, Long addressId, AddressDto addressDto) {
+    public boolean updateAddress(Long userId, Long addressId, AddressDto addressDto) {
+        if (!userRepository.isUserInDatabase(userId) || !addressRepository.isAddressInDatabase(userId, addressId)) {
+            return false;
+        }
         Address address = addressDtoMapper.addressDtoToAddress(addressDto);
         addressRepository.deleteAddress(userId, addressId);
         addressRepository.save(userId, address);
+        return true;
     }
 
-    public void deleteAddress(Long userId, Long addressId) {
-        addressRepository.deleteAddress(userId, addressId);
+    public boolean deleteAddress(Long userId, Long addressId) {
+        if (!userRepository.isUserInDatabase(userId)) {
+            return false;
+        } else {
+            addressRepository.deleteAddress(userId, addressId);
+            return true;
+        }
     }
 
     public List<AddressDto> getAddresses(Long userId) {
+        if (addressRepository.findAllByUserId(userId) == null) {
+            return null;
+        }
         return addressRepository.findAllByUserId(userId)
                 .stream()
                 .map(address -> {
-                    var user = userRepository.getUser(address.getUserId());
-                    return addressDtoMapper.addressToAddressDto(address, user);
+                    Optional<User> optionalUser = userRepository.getUser(address.getUserId());
+                    if (optionalUser.isPresent()) {
+                        return addressDtoMapper.addressToAddressDto(address, optionalUser);
+                    } else {
+                        throw new RuntimeException("User not found");
+                    }
                 })
+                .filter(Objects::nonNull)
                 .collect(Collectors.toList());
     }
